backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
pandas==2.0.3
openpyxl==3.1.2
SQLAlchemy==2.0.19
mysql-connector-python==8.1.0
python-dotenv==1.0.0
google-generativeai==0.3.0
numpy==1.24.3
Werkzeug==2.3.6